include/bigquery_cdc/requirements.txt
google-cloud-bigquery-storage==2.26.0
google-cloud-bigquery==3.25.0